backend/requirements.txt
asgiref==3.3
certifi==2020.4.5.1
chardet==3.0.4
django==3.0.5
djangorestframework==3.11.0
idna==2.9
pyparsing==2.4.7
pytest-django==3.9.0
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
django-filter~=2.4.0
PyJWT~=2.1.0
djangorestframework-simplejwt==4.7.2
psycopg2-binary==2.8.6
gunicorn==20.0.4
cffi==1.14.6
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
diff-match-patch==20200713
django-colorfield==0.4.2
django-import-export==2.5.0
django-templated-mail==1.1.1
djoser==2.1.0
drf-extra-fields==3.1.1
drf-writable-nested==0.6.3
et-xmlfile==1.1.0
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.1
MarkupPy==1.14
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.3.1
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
python3-openid==3.2.0
PyYAML==5.4.1
requests-oauthlib==1.3.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
tablib==3.0.0
typing-extensions==3.10.0.0
uritemplate==3.0.1
xlrd==2.0.1
xlwt==1.3.0
